+++ testdata/de/dm/intellij/liferay/index/PortletIndexTest/de/dm/portlet/UnnamedPortlet.java
@@ -1,0 +1,10 @@
+package de.dm.portlet;
+
+import org.osgi.service.component.annotations.Component;
+import javax.portlet.Portlet;
+
+@Component(
+    service = Portlet.class
+)
+public class UnnamedPortlet implements Portlet {
+}