+++ changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Bugfix: Correctly detect locale strings in JSON schema for journal structure files
 - Bugfix: Fixed compatibility with IntelliJ 2018.3 when injecting javascript into freemarker taglibs
 - Feature: Code completion for portlet names in JSP taglibs
+- Feature: Support for multiple OSGi service definitions in a @Component annotation
 
 ## 0.0.3
 - Feature: Javascript injection for aui:validator and many onXXX attributes